created file structure
