+++ models/model.go
@@ -1,0 +1,7 @@
+package models
+
+type Article struct {
+	Title   string `json:"title"`
+	Desc    string `json:"desc"`
+	Content string `json:"content"`
+}
